Add ability to step through rounds with arrow keys. There seems to be a bug somewhere causing a bias in the weights.

+++ src/Game.h
@@ -17,7 +17,7 @@ namespace cagesim
     {
         // per round, per player, per strategy
         std::vector<std::vector<std::vector<float>>> strategyCosts;
-        std::vector<std::vector<std::vector<float>>> strategyWeights;
+        std::vector<std::vector<std::vector<double>>> strategyWeights;
 
         void AddRound(size_t numPlayers, size_t numStrategies);
     };
