+++ src/Strategy.h
@@ -15,7 +15,7 @@ namespace cagesim
         Strategy(size_t nPlayers) { numPlayers = nPlayers; }
         size_t GetNumPlayers() const { return numPlayers; }
         size_t GetNumStrategies() const { return numStrategies; }
-        virtual std::function<float(size_t, size_t, const std::vector<size_t>&)> GetCostFunction() const = 0;
+        virtual std::function<float(uint, uint, const std::vector<uint>&)> GetCostFunction() const = 0;
         virtual void AddData(const GameData& d) = 0;
     protected:
         size_t numPlayers;
